% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressions.R
\name{AKM_construct_share_matrix}
\alias{AKM_construct_share_matrix}
\title{Construct share matrix for AKM estimation}
\usage{
AKM_construct_share_matrix(industry_location_data)
}
\arguments{
\item{industry_location_data}{Industry location data}
}
\description{
Construct share matrix for AKM estimation
}
