% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressions.R
\name{AKM_location_regressions}
\alias{AKM_location_regressions}
\title{Location regressions}
\usage{
AKM_location_regressions(
  location_data,
  AKM_share_matrix,
  industry_clusters = NULL
)
}
\arguments{
\item{location_data}{Location data}

\item{AKM_share_matrix}{AKM share matrix}

\item{industry_clusters}{(optional) default is NULL}
}
\description{
Location regressions
}
