% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressions.R
\name{BHJ_check_proposition1}
\alias{BHJ_check_proposition1}
\title{Check BHJ proposition 1}
\usage{
BHJ_check_proposition1(results)
}
\arguments{
\item{results}{Results}
}
\description{
Check BHJ proposition 1
}
