% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressions.R
\name{BHJ_industry_regressions}
\alias{BHJ_industry_regressions}
\title{BHJ industry regressions}
\usage{
BHJ_industry_regressions(
  industry_data,
  lagIV = FALSE,
  lagsIV = 1,
  timeFE = FALSE,
  industryFE = FALSE
)
}
\arguments{
\item{industry_data}{Industry data}

\item{lagIV}{(optional) default is FALSE}

\item{timeFE}{(optional) default is FALSE}

\item{industryFE}{(optional) default is FALSE}
}
\description{
BHJ industry regressions
}
