% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressions.R
\name{BHJ_transform_to_industry}
\alias{BHJ_transform_to_industry}
\title{Apply the BHJ transformation to the industry level}
\usage{
BHJ_transform_to_industry(
  industry_location_data,
  lagIV = FALSE,
  lagsIV = 1,
  industryFE = FALSE,
  timeFE = FALSE
)
}
\arguments{
\item{industry_location_data}{Industry location data}

\item{lagIV}{(optional) default is FALSE}

\item{industryFE}{(optional) default is FALSE}

\item{timeFE}{(optional) default is FALSE}
}
\description{
Apply the BHJ transformation to the industry level
}
