% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{SSIV_check_shares}
\alias{SSIV_check_shares}
\title{Check the industry_location_shares}
\usage{
SSIV_check_shares(industry_location_shares)
}
\arguments{
\item{industry_location_shares}{Industry location shares}
}
\description{
Check the industry_location_shares
}
