% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{SSIV_checks}
\alias{SSIV_checks}
\title{Check that data provided meets the BHJ structure}
\usage{
SSIV_checks(
  location_outcomes,
  industry_shocks,
  industry_location_shares,
  lagIV = FALSE
)
}
\arguments{
\item{location_outcomes}{Location outcomes}

\item{industry_shocks}{Industry shocks}

\item{industry_location_shares}{Industry location shares}
}
\description{
Check that data provided meets the BHJ structure
}
