% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{SSIV_outcomes}
\alias{SSIV_outcomes}
\title{This function has two tasks: residualize Y,X variables, then collapse to location}
\usage{
SSIV_outcomes(
  location_outcomes,
  time_variable = "",
  unit_identifier = "location",
  Y_vars,
  X_vars,
  W_var,
  linear_covariates = c(),
  FE_covariates = c()
)
}
\arguments{
\item{location_outcomes}{Location outcomes}

\item{time_variable}{(optional) panel dimension}

\item{unit_identifier}{(optional) unique unit identifier, assumed to be 'location' if not specific}

\item{Y_vars}{vector of outcome variable names (missings are allowed)}

\item{X_vars}{vector of X variable names (missings are allowed but discouraged)}

\item{W_var}{name of weights}

\item{linear_covariates}{(optional) default is empty}

\item{FE_covariates}{(optional) default is empty}
}
\description{
This function has two tasks: residualize Y,X variables, then collapse to location
}
