% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{SSIV_panelID}
\alias{SSIV_panelID}
\title{Replace industry with (industry,date) and location with (location,date)}
\usage{
SSIV_panelID(
  location_outcomes,
  industry_shocks,
  industry_location_shares,
  time_variable
)
}
\arguments{
\item{location_outcomes}{Location outcomes}

\item{industry_shocks}{Industry shocks}

\item{industry_location_shares}{Industry location shares}

\item{time_variable}{Time identifier}
}
\description{
Replace industry with (industry,date) and location with (location,date)
}
