% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{SSIV_residualization_formula}
\alias{SSIV_residualization_formula}
\title{Define the residualization formula.}
\usage{
SSIV_residualization_formula(linear_covariates = c(), FE_covariates = c())
}
\arguments{
\item{linear_covariates}{(optional) default is empty}

\item{FE_covariates}{(optional) default is empty}
}
\description{
Define the residualization formula.
}
