% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{SSIV_residualization_with_missings}
\alias{SSIV_residualization_with_missings}
\title{This function has two tasks: residualize Y,X variables, then collapse to location}
\usage{
SSIV_residualization_with_missings(
  dd,
  Y_var,
  W_var,
  time_variable,
  unit_identifier,
  linear_covariates = c(),
  FE_covariates = c()
)
}
\arguments{
\item{dd}{Data in which to replace variables with residualized versions.}

\item{Y_var}{outcome variable (missings are allowed)}

\item{W_var}{name of weights}

\item{time_variable}{(optional) panel dimension}

\item{unit_identifier}{(optional) unique unit identifier, assumed to be 'location' if not specific}

\item{linear_covariates}{(optional) default is empty}

\item{FE_covariates}{(optional) default is empty}
}
\description{
This function has two tasks: residualize Y,X variables, then collapse to location
}
