% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressions.R
\name{Uncorrected_location_regressions}
\alias{Uncorrected_location_regressions}
\title{Raw location regressions}
\usage{
Uncorrected_location_regressions(location_data, lagIV = FALSE)
}
\arguments{
\item{location_data}{Location data}

\item{lagIV}{(optional) default is FALSE}
}
\description{
Raw location regressions
}
